Add json files to 1

+++ app/Http/Controllers/ProductController.php
@@ -171,17 +171,18 @@ class ProductController extends Controller
     }
         public function delete_product($id)
     {
-        $product=product::with('files')->where('id',$id)->first();
-        if ($product->files) {
-            foreach ($product->files as $file) {
-            if (Storage::disk('public')->exists($file->file_path)) {
-                Storage::disk('public')->delete($file->file_path);
+        $product = product::with('files')->where('id', $id)->first();
+        if ($product) {
+            if ($product->files && $product->files->count() > 0) {
+                foreach ($product->files as $file) {
+                    if (Storage::disk('public')->exists($file->file_path)) {
+                        Storage::disk('public')->delete($file->file_path);
+                    }
+                    $file->delete();
+                }
             }
-            $file->delete();
+            $product->delete();
         }
-        }
-        $product->delete();
-
         return back();
     }
 }
